refactor(commands): return typed results instead of printing directly
- Add commands/types.rs with ReposResult, IssuesResult, TriageResult
- Add output.rs with centralized rendering for text/json/yaml formats
- Commands return data; mod.rs handles spinners and rendering dispatch
- Move helper functions (truncate_title, format_relative_time) to output.rs
- Split triage into analyze() and post() for proper confirmation flow
- Fix: render triage output before asking for confirmation
- Replace unwrap() with expect() for better error messages
- Improves testability and prepares for future library extraction
